requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
numpy>=1.24.0

# LangChain integration
langchain>=0.1.0
langchain-community>=0.0.10

# RL utilities
gymnasium>=0.29.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Logging and monitoring
tensorboard>=2.15.0
tqdm>=4.66.0

# Vector store for memory
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Utilities
dataclasses-json>=0.6.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code execution safety
RestrictedPython>=6.0

# Math evaluation
sympy>=1.12

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3
